#67 Split persistence.ts into helpers, quiz, user and prisma components
